+++ NEWS.md
@@ -1,4 +1,5 @@
 # tna 1.0.0
+
 * CRAN release
 
 # tna 0.7.0